requirements.txt
streamlit
mysql-connector-python
langchain
langchain-community
groq
python-dotenv
sentence-transformers
faiss-cpu
langchain-groq
langchain-huggingface